Prevent readonly parameters from being rebound
According so S06, "A [readonly] parameter may also not be rebound;
trying to do so results in a compile time error."  However, this
behavior is not currently tested in Roast, and is inconsistently
enforced by Rakudo.

Specifically, prior to this commit, the following code generated a
compile-time X::Bind error

    sub f($a) { $a := 0 }

but neither of the following two lines did:

    sub f(Mu $a) { $a := 0   }
    sub g(   @a) { @a := [0] }

This commit fixes that issue by calling the existing
is_lexical_marked_ro method to determine whether a parameter is
readonly.  It also annotates parameters as readonly or not readonly in
a few locations (which is necessary because is_lexical_marked_ro is
used before the Parameter objects are constructed).

Finally, this commit changes the one location where Rakudo code was
taking advantage of the previous bug by rebinding a readonly
parameter.

Because signature destructuring operates by binding the parameters in
a signature, this commit also fixes the incorrectly-allowed rebinding
of variables with values set in that way.

Both for functions and standalone signatures, non-readonly
parameters (that is, any that are `copy` or `rw`) can be freely
rebound. Thus,

    my ($a is copy, $b) := (0, 1); $a := 42;

is correctly allowed, but generates a compilation error without the
`is copy`.